feat: Implement various UI and routing updates
- Adds a 'Back to Home' link on the /thank-you page to make navigation easier for you.
- Adds a support email link (support@oneappperday.com) to the header of the homepage.
- Refactors the quiz to use unique URLs for each step (e.g., /quiz/1, /quiz/2). This improves navigation and allows you to use the back button.
- The quiz state (answers) is now persisted in localStorage to be maintained across steps.
- The final payment screen's URL and structure remain unchanged as you requested.
